ch02/pyproject.toml
[project]
name = "ch02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.73.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "jupyterlab>=4.5.7",
    "python-dotenv>=1.2.2",
]
